✨ feat(spring-security-core): 删除动态安全过滤器类，添加权限拦截器和Web过滤器类
✨ feat(spring-business): 添加权限相关请求DTO，控制器接口及实现，服务层实现

✨ feat(spring-security-api): 修改权限角色引用及用户权限引用实体，调整权限ID类型为String

✨ feat(spring-business): 修改字典控制器，增加权限检查逻辑

✨ feat(spring-security-core): 添加权限配置类，支持权限设置

✨ feat(spring-business): 添加权限映射DTO类

✨ feat(spring-business): 添加权限用户分页响应DTO类

✨ feat(spring-business): 添加权限角色分页请求DTO类

✨ feat(spring-business): 添加权限用户添加请求DTO类

✨ feat(spring-business): 添加权限角色添加请求DTO类

✨ feat(spring-business): 添加权限用户分页请求DTO类

✨ feat(spring-business): 添加权限角色更新请求DTO类

✨ feat(spring-security-core): 修改构建脚本，添加Spring WebFlux依赖

✨ feat(spring-business): 修改字典添加请求DTO，增加非空校验

✨ feat(spring-security-api): 修改用户实体，添加密码过期时间字段

✨ feat(spring-business): 添加Servlet环境下的权限控制器实现

✨ feat(spring-business): 添加WebFlux环境下的权限控制器实现

✨ feat(spring-security-core): 添加抽象权限类，提供权限提取和检查逻辑

+++ spring-business/src/main/java/org/start2do/controller/servlet/SysDictController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.start2do.dto.BusinessException;
 import org.start2do.dto.IdStrReq;
 import org.start2do.dto.Page;
 import org.start2do.dto.R;
@@ -32,7 +33,7 @@ import org.start2do.util.BeanValidatorUtil;
 @RequiredArgsConstructor
 @RequestMapping("/dict")
 @ConditionalOnWebApplication(type = Type.SERVLET)
-@ConditionalOnProperty(prefix = "start2do.business.controller", name = "dict", havingValue = "true",matchIfMissing = true)
+@ConditionalOnProperty(prefix = "start2do.business.controller", name = "dict", havingValue = "true", matchIfMissing = true)
 public class SysDictController {
 
     private final SysDictService sysDictService;
@@ -55,6 +56,10 @@ public class SysDictController {
     @SysLogSetting("删除字典")
     public R delete(IdStrReq req) {
         BeanValidatorUtil.validate(req);
+        SysDict dict = sysDictService.findOneById(req.getId());
+        if (dict.getDictType() == SysDict.Type.SYSTEM) {
+            throw new BusinessException("系统内置无法修改");
+        }
         sysDictService.remove(req.getId());
         return R.ok();
     }
@@ -67,6 +72,9 @@ public class SysDictController {
     public R add(@RequestBody DictAddReq req) {
         BeanValidatorUtil.validate(req);
         SysDict sysDict = DictDtoMapper.INSTANCE.toSysDict(req);
+        if (sysDictService.exists(new QSysDict().dictName.eq(sysDict.getDictName()))) {
+            throw new BusinessException("已被使用");
+        }
         sysDictService.save(sysDict);
         return R.ok();
     }
@@ -79,6 +87,11 @@ public class SysDictController {
     public R update(@RequestBody DictUpdateReq req) {
         BeanValidatorUtil.validate(req);
         SysDict dict = sysDictService.getById(req.getId());
+        if (!dict.getDictName().equals(req.getDictName())) {
+            if (sysDictService.exists(new QSysDict().dictName.eq(req.getDictName()))) {
+                throw new BusinessException("已被使用");
+            }
+        }
         DictDtoMapper.INSTANCE.updateSysDict(dict, req);
         sysDictService.update(dict);
         return R.ok();
